requirements.txt
locust==2.32.4
python-dotenv==1.0.1
